+++ scenarios/NGSIM/peachtree/scenario.py
@@ -1,18 +1,18 @@
 from pathlib import Path
 
-from smarts.sstudio import gen_scenario
 from smarts.sstudio import types as t
+from smarts.sstudio import gen_scenario
 
 
 traffic_histories = [
     t.TrafficHistoryDataset(
-        name=f"us101_{hd}",
+        name=f"peach_{hd}",
         source_type="NGSIM",
         input_path=None,  # for example: f"./trajectories-{hd}.txt"
         speed_limit_mps=28,
         default_heading=0,
     )
-    for hd in ["0750am-0805am", "0805am-0820am", "0820am-0835am"]
+    for hd in ["0400pm-0415pm", "1245pm-0100pm"]
 ]
 
 gen_scenario(